Removing structures that are not required anymore. Adding data types that will be used accross rdbms and nosql model service.

+++ src/SoundFingerprinting/Data/ISubFingerprintReference.cs
@@ -2,6 +2,6 @@
 {
     public interface ISubFingerprintReference
     {
-        long HashCode { get; }
+        int HashCode { get; }
     }
 }
